pyproject.toml
[project]
name = "fastapi-auth-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Gucik",email = "danielgucik7@gmail.com"}
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "bcrypt (==4.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.9,<0.15.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "pre-commit (>=4.5.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

# =========================================================
# Linter & Formatter
# =========================================================

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "UP", "B", "S"]
extend-ignore = ["D100", "D104", "D105", "D106", "D107"]
src = ["src"]
exclude = ["alembic"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["auth", "config", "shared"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
mypy_path = ["src", "src/modules"]
namespace_packages = true
explicit_package_bases = true


[[tool.mypy.overrides]]
module = ["passlib.*", "jose.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
